Lelijkste type cast ooit!
Schiet nu zichzelf dood

+++ Statements/worms/model/programs/parser/FireStatement.java
@@ -13,7 +13,7 @@ public class FireStatement extends Statement {
 
 	@Override
 	public void execute(Worm activeWorm, IActionHandler handler) {
-		handler.fire(activeWorm, (int) this.yieldExpression.getValue(activeWorm));
+		handler.fire(activeWorm, (int) ((double) this.yieldExpression.getValue(activeWorm)));
 	}
